# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: API de Produtos
  #  title: PEGASUS-SA → API de Produtos
  termsOfService: http://localhost:8080/terms
  description: |
    API RESTful CRUD de controle de produtos para estudo.  
    
    #### 🍵🍵🍵 NOTAS DO DESENVOLVEDOR 🍵🍵🍵    
    
    ##### Metricas
    - Todas as métricas disponíveis: `http://localhost:8080/actuator/metrics`
    - Métrica customizada de contagem de requests: `http://localhost:8080/actuator/metrics/http_requests_total`
    - Métrica customizada de duração de requests: `http://localhost:8080/actuator/metrics/http_request_duration_ms`
    - Formato Prometheus (todas as métricas): ` http://localhost:8080/actuator/prometheus`
    
    ##### DataBase `http://localhost:8080/h2-console`
    - Driver class: `ver application.yaml` → `spring.datasource.driver-class-name`
    - JDBC URL: `ver application.yaml` → `spring.datasource.url`
    - User Name: `ver application.yaml` → `spring.datasource.username`
    - Password: `ver application.yaml` → `spring.datasource.password`


  contact:
    name: Enterprise
    email: suporte@pegasus.com.br
    url: http://localhost:8080
  license:
    name: License
    url: http://localhost:8080/license
servers:
  - url: http://localhost:8080
    description: Servidor local para solicitações de produtos

tags:
  - { name: product, description: produto }

paths:

  /product:
    parameters:
      - $ref: '#/components/parameters/_header_x_trace_id'

    get:
      tags: [ product ]
      operationId: find-all-product
      parameters:
        - $ref: '#/components/parameters/_query_page'
        - $ref: '#/components/parameters/_query_size'
      summary: Busca uma paginação de elementos
      responses:
        '200':
          description: Retorna a paginação solicitada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/product-page-response-type' }
        "400": { $ref: '#/components/responses/bad-request-response' }
        "500": { $ref: '#/components/responses/internal-server-error-response' }

    post:
      tags: [ product ]
      operationId: create-product
      summary: Cria um novo elemento
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/product-create-body-type' }
      responses:
        '201':
          description: Retorna o elemento criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/product-type' }
        "400": { $ref: '#/components/responses/bad-request-response' }
        "409": { $ref: '#/components/responses/conflict-response' }
        "500": { $ref: '#/components/responses/internal-server-error-response' }

  /product/{id}:
    parameters:
      - $ref: '#/components/parameters/_header_x_trace_id'
      - $ref: '#/components/parameters/_path_id'

    get:
      tags: [ product ]
      operationId: find-by-id-product
      summary: Busca um elemento pelo ID
      responses:
        '200':
          description: O elemento correspondente ao ID
          content:
            application/json:
              schema: { $ref: '#/components/schemas/product-type' }
        "404": { $ref: '#/components/responses/not-found-response' }
        "500": { $ref: '#/components/responses/internal-server-error-response' }

    put:
      tags: [ product ]
      operationId: update-product
      summary: Atualiza um elemento pelo ID (por completo)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/product-update-body-type' }
      responses:
        '200': { description: ok }
        "400": { $ref: '#/components/responses/bad-request-response' }
        "404": { $ref: '#/components/responses/not-found-response' }
        "409": { $ref: '#/components/responses/conflict-response' }
        "500": { $ref: '#/components/responses/internal-server-error-response' }

    delete:
      tags: [ product ]
      operationId: delete-product
      summary: Deleta um elemento pelo ID
      responses:
        '204': { description: ok }
        "404": { $ref: '#/components/responses/not-found-response' }
        "500": { $ref: '#/components/responses/internal-server-error-response' }

components:
  schemas:
    att-product-name-create-update: { type: string, minLength: 1, example: Chocolate, description: Nome do produto. Obrigatório e não pode ser vazio }
    att-product-price-create-update: { type: number, format: float, minimum: 0,multipleOf: 0.01, example: 3.52, description: Preço do produto. Obrigatório e não pode ser negativo. }
    att-product-quantity-create-update: { type: integer, minimum: 0, example: 10, description: Quantidade do produto. Obrigatório e não pode ser negativo. }

    product-type:
      type: object
      properties:
        id: { type: integer }
        name: { $ref: '#/components/schemas/att-product-name-create-update' }
        price: { $ref: '#/components/schemas/att-product-price-create-update' }
        quantity: { $ref: '#/components/schemas/att-product-quantity-create-update' }

    product-page-response-type:
      type: object
      properties:
        pagination: { $ref: '#/components/schemas/pagination-type' }
        data:
          type: array
          items: { $ref: '#/components/schemas/product-type' }

    product-create-body-type:
      type: object
      required: [ name, price, quantity ]
      properties:
        name: { $ref: '#/components/schemas/att-product-name-create-update' }
        price: { $ref: '#/components/schemas/att-product-price-create-update' }
        quantity: { $ref: '#/components/schemas/att-product-quantity-create-update' }

    product-update-body-type:
      type: object
      required: [ name, price, quantity ]
      properties:
        name: { $ref: '#/components/schemas/att-product-name-create-update' }
        price: { $ref: '#/components/schemas/att-product-price-create-update' }
        quantity: { $ref: '#/components/schemas/att-product-quantity-create-update' }

    pagination-type:
      type: object
      properties:
        page: { type: integer, example: 10, description: Página atual (0-based) }
        size: { type: integer, example: 30, description: Quantidade de elementos por página }
        elements: { type: integer, format: int64, example: 300, description: Quantidade de elementos retornados na página }
        pages: { type: integer, example: 10, description: Quantidade de páginas disponíveis }
        previous: { type: boolean, example: true, description: Verificação de existência de página anterior }
        next: { type: boolean, example: true, description: Verificação de existência de página posterior }

    exception-response-type:
      type: object
      description: Estrutura padrão de erro da API
      required: [ status, error, message ]
      properties:
        traceId: { type: string, example: a12f3cde-45b6-7890-abc1-def234567890, description: TraceId enviado na request }
        status: { type: integer, format: int32, example: 400 }
        error: { type: string, example: "Bad Request", description: "Descrição curta do erro HTTP" }
        message: { type: string, example: "Validation error", description: "Mensagem explicando o erro de forma geral" }
        timestamp: { type: string, format: date-time, example: "2025-11-01T12:34:56Z", description: "Data/hora do erro em ISO-8601 (UTC recomendado)" }
        path: { type: string, example: "/product", description: "URI do endpoint requisitado" }

  responses:

    bad-request-response:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/exception-response-type'
          example:
            - code: 400
              message: Bad Request
              messageDetail: Invalid Parameter X

    not-found-response:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/exception-response-type'
          example:
            - code: 404
              message: Not Found
              messageDetail: Element not found

    conflict-response:
      description: Elemento já Cadastrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/exception-response-type'
          example:
            - code: 409
              message: Conflict
              messageDetail: Element already registered

    internal-server-error-response:
      description: Erro interno no servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/exception-response-type'
          example:
            - code: 500
              message: Internal Server Error
              messageDetail: PROBLEMS IN QUERYING THE DATABASE

  parameters:
    _header_x_trace_id:
      name: X-Trace-Id
      in: header
      description: Id da requisição
      required: true
      schema:
        type: string
        format: uuid
        example: "a12f3cde-45b6-7890-abc1-def234567890"

    _query_page:
      description: Não pode ser negativo
      in: query
      name: page
      schema:
        type: integer
        example: 0
        default: 0

    _query_size:
      description: Não pode ser negativo
      in: query
      name: size
      schema:
        type: integer
        example: 10
        default: 10

    _path_id:
      description: Não pode ser nulo, negativo ou em branco
      in: path
      name: id
      required: true
      schema:
        type: integer